requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
pdfkit==1.0.0
openai==1.50.0